feat: post 작성 기능과 연관된 repository 생성

+++ backend/src/main/java/com/kkini/core/domain/post/dto/request/PostRegisterRequestDto.java
@@ -18,7 +18,7 @@ public class PostRegisterRequestDto {
     private String contents;
 
     @Schema(description = "레시피")
-    private int recipeId;
+    private Long recipeId;
 
     @Schema(description = "이미지")
     private List<MultipartFile> images;
